quarto_inspect {quarto}	R Documentation

Inspect Quarto Input File or Project

Description

Inspect a Quarto project or input path. Inspecting a project returns its config and engines. Inspecting an input path return its formats, engine, and dependent resources.

Usage

quarto_inspect(input = ".")

Arguments

input
The input file or project directory to inspect.

Value

Named list. For input files, the list has members engine, format, and resources. For projects the list has members engines and config

Examples

## Not run: 
# Inspect input file file
quarto_inspect("notebook.Rmd")

# Inspect project
quarto_inspect("myproject")

## End(Not run)

[Package quarto version 1.3 Index]