quarto_version {quarto}	R Documentation

Check quarto version

Description

Determine the specific version of quartobinary found by quarto_path(). If it returns 99.9.9 then it means you are using a dev version.

Usage

quarto_version()

Value

a numeric_version with the quarto version found

[Package quarto version 1.3 Index]