Create a quarto project

This function calls quarto create project <type> <name>. It will create a new directory with the project name and add some skeletons files for the project type chosen.

Usage

quarto_create_project(
  name,
  type = "default",
  dir = ".",
  no_prompt = FALSE,
  quiet = FALSE,
  quarto_args = NULL
)

Arguments

name

The name of the project and the directory that will be created.

type

The type of project to create. As of 1.4, it can be one of default, website, blog, book, manuscript, confluence.

dir

The directory where to create the new Quarto project.

no_prompt

Do not prompt to confirm approval to download external extension.

quiet

Suppress warning and other messages.

quarto_args

Character vector of other quarto CLI flag pass to the command. This is mainly for advanced usage, e.g it can be useful for new options added to quarto CLI and not yet supported as function argument.

Quarto version required

This function require Quarto 1.4 or higher. Use quarto_version()to check the version of Quarto detected.

See also