quarto_render {quarto}	R Documentation

Render Markdown

Description

Render the input file to the specified output format using quarto. If the input requires computations (e.g. for Rmd or Jupyter files) then those computations are performed before rendering.

Usage

quarto_render(
  input = NULL,
  output_format = NULL,
  output_file = NULL,
  execute = TRUE,
  execute_params = NULL,
  execute_dir = NULL,
  execute_daemon = NULL,
  execute_daemon_restart = FALSE,
  execute_debug = FALSE,
  use_freezer = FALSE,
  cache = NULL,
  cache_refresh = FALSE,
  debug = FALSE,
  quiet = FALSE,
  pandoc_args = NULL,
  as_job = getOption("quarto.render_as_job", "auto")
)

Arguments

input
The input file or project directory to be rendered (defaults to rendering the project in the current working directory).

output_format
Target output format (defaults to "html"). The option "all" will render all formats defined within the file or project.

output_file
The name of the output file. If using NULL then the output filename will be based on filename for the input file.

execute
Whether to execute embedded code chunks.

execute_params
A list of named parameters that override custom params specified within the YAML front-matter.

execute_dir
The working directory in which to execute embedded code chunks.

execute_daemon
Keep Jupyter kernel alive (defaults to 300 seconds). Note this option is only applicable for rendering Jupyter notebooks or Jupyter markdown.

execute_daemon_restart
Restart keepalive Jupyter kernel before render. Note this option is only applicable for rendering Jupyter notebooks or Jupyter markdown.

execute_debug
Show debug output for Jupyter kernel.

use_freezer
Force use of frozen computations for an incremental file render.

cache
Cache execution output (uses knitr cache and jupyter-cache respectively for Rmd and Jupyter input files).

cache_refresh
Force refresh of execution cache.

debug
Leave intermediate files in place after render.

quiet
Suppress warning and other messages.

pandoc_args
Additional command line options to pass to pandoc.

as_job
Render as an RStudio background job. Default is "auto", which will render individual documents normally and projects as background jobs. Use the quarto.render_as_job R option to control the default globally.

Examples

## Not run: 
# Render R Markdown
quarto_render("notebook.Rmd")
quarto_render("notebook.Rmd", output_format = "pdf")

# Render Jupyter Notebook
quarto_render("notebook.ipynb")

# Render Jupyter Markdown
quarto_render("notebook.md")

## End(Not run)

[Package quarto version 1.3 Index]